create project

+++ src/main/java/edu/ijjse/dehivalazoomanagemetsystem/controller/AnimalMngController.java
@@ -1,10 +1,12 @@
 package edu.ijjse.dehivalazoomanagemetsystem.controller;
 
 import com.jfoenix.controls.JFXButton;
+import com.jfoenix.controls.JFXComboBox;
 import com.jfoenix.controls.JFXTextField;
 import edu.ijjse.dehivalazoomanagemetsystem.dto.AnimalMngDto;
 import edu.ijjse.dehivalazoomanagemetsystem.dto.tm.AnimalMngTM;
 import edu.ijjse.dehivalazoomanagemetsystem.model.AnimalMngModel;
+import edu.ijjse.dehivalazoomanagemetsystem.utill.RegexUtill;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -14,6 +16,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -25,6 +28,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class AnimalMngController implements Initializable {
@@ -45,13 +49,41 @@ public class AnimalMngController implements Initializable {
      try {
          tabelload();
          nextAnimalId();
+         getEnclosureId();
+         getdivisionIds();
 
      } catch (SQLException e) {
          new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
      }
  }
 
- private void tabelload() throws SQLException {
+    private void getdivisionIds() {
+        try {
+            ArrayList<String> division = model.getDivision();
+            ObservableList<String> objects = FXCollections.observableArrayList();
+            objects.addAll(division);
+            devisiontxt.setItems(objects);
+        } catch (SQLException e) {
+            new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
+            e.printStackTrace();
+        }
+
+
+    }
+
+    private void getEnclosureId() {
+        try {
+            ArrayList<String> enclosure = model.getEnclosure();
+            ObservableList<String> objects = FXCollections.observableArrayList();
+            objects.addAll(enclosure);
+            enclosrtxt.setItems(objects);
+        } catch (SQLException e) {
+            new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
+            e.printStackTrace();
+        }
+    }
+
+    private void tabelload() throws SQLException {
   ArrayList<AnimalMngDto> animalMngDtos = model.getAll();
   ObservableList<AnimalMngTM> AnimalTms = FXCollections.observableArrayList();
 for (AnimalMngDto animalMngDto : animalMngDtos) {
@@ -90,17 +122,16 @@ animaltbl.setItems(AnimalTms);
  @FXML
  private JFXButton deletebtn;
 
- @FXML
- private JFXTextField devisiontxt;
+    @FXML
+    private JFXComboBox<String> devisiontxt;
 
  @FXML
  private TableColumn<AnimalMngTM, String> devisoncol;
 
  @FXML
  private TableColumn<AnimalMngTM, String> enclosercol;
-
- @FXML
- private JFXTextField enclosrtxt;
+    @FXML
+    private JFXComboBox<String> enclosrtxt;
 
  @FXML
  private JFXTextField gemdertxt;
@@ -137,24 +168,43 @@ animaltbl.setItems(AnimalTms);
         double age = Double.parseDouble(agetxt.getText());
         String gender = gemdertxt.getText();
         String catogary = catogarytxt.getText();
-        String encloser = enclosrtxt.getText();
-        String devision = devisiontxt.getText();
+        String encloser = enclosrtxt.getValue();
+        String devision = devisiontxt.getValue();
 
+        idtxt.setStyle(idtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        nametxt.setStyle(nametxt.getStyle() + ";-fx-border-color: #7367F0;");
+        agetxt.setStyle(agetxt.getStyle() + ";-fx-border-color: #7367F0;");
+        gemdertxt.setStyle(gemdertxt.getStyle() + ";-fx-border-color: #7367F0;");
+        catogarytxt.setStyle(catogarytxt.getStyle() + ";-fx-border-color: #7367F0;");
+        enclosrtxt.setStyle(enclosrtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        devisiontxt.setStyle(devisiontxt.getStyle() + ";-fx-border-color: #7367F0;");
 
-        AnimalMngDto dto = new AnimalMngDto(animalId, name, age,gender, catogary, encloser, devision);
-        try {
-            boolean update = model.update(dto);
-            if (update) {
-                new Alert(Alert.AlertType.INFORMATION, "Animal Updated Successfully").show();
-                clearform();
-               tabelload();
-               nextAnimalId();
-            }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR, e.toString()).show();
-            e.printStackTrace();
+        boolean isvalidAge = RegexUtill.IsvalidAge(agetxt.getText());
+        boolean isValidGender = RegexUtill.IsValidGender(gemdertxt.getText());
+
+        if (!isvalidAge){
+            agetxt.setStyle(agetxt.getStyle() + ";-fx-border-color: red;");
+            System.out.println("not a valid age");
         }
-
+        if (!isValidGender){
+            gemdertxt.setStyle(gemdertxt.getStyle() + ";-fx-border-color: red;");
+            return;
+        }
+        if (isValidGender&&isvalidAge){
+            AnimalMngDto dto = new AnimalMngDto(animalId, name, age,gender, catogary, encloser, devision);
+            try {
+                boolean update = model.update(dto);
+                if (update) {
+                    new Alert(Alert.AlertType.INFORMATION, "Animal Updated Successfully").show();
+                    clearform();
+                    tabelload();
+                    nextAnimalId();
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, e.toString()).show();
+                e.printStackTrace();
+            }
+        }
     }
 
     @FXML
@@ -164,22 +214,43 @@ animaltbl.setItems(AnimalTms);
      double age = Double.parseDouble(agetxt.getText());
      String gender = gemdertxt.getText();
      String catogary = catogarytxt.getText();
-     String encloser = enclosrtxt.getText();
-     String devision = devisiontxt.getText();
+     String encloser = enclosrtxt.getValue();
+     String devision = devisiontxt.getValue();
 
+        idtxt.setStyle(idtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        nametxt.setStyle(nametxt.getStyle() + ";-fx-border-color: #7367F0;");
+        agetxt.setStyle(agetxt.getStyle() + ";-fx-border-color: #7367F0;");
+        gemdertxt.setStyle(gemdertxt.getStyle() + ";-fx-border-color: #7367F0;");
+        catogarytxt.setStyle(catogarytxt.getStyle() + ";-fx-border-color: #7367F0;");
+        enclosrtxt.setStyle(enclosrtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        devisiontxt.setStyle(devisiontxt.getStyle() + ";-fx-border-color: #7367F0;");
 
-     AnimalMngDto dto = new AnimalMngDto(animalId, name, age,gender, catogary, encloser, devision);
-        try {
-            boolean add = model.add(dto);
-            if (add) {
-                new Alert(Alert.AlertType.CONFIRMATION, "Animal Added Successfully").show();
-                clearform();
-                tabelload();
-                nextAnimalId();
+        boolean isvalidAge = RegexUtill.IsvalidAge(agetxt.getText());
+        boolean isValidGender = RegexUtill.IsValidGender(gemdertxt.getText());
+
+        if (!isvalidAge){
+            agetxt.setStyle(agetxt.getStyle() + ";-fx-border-color: red;");
+            System.out.println("not a valid age");
+        }
+        if (!isValidGender){
+            gemdertxt.setStyle(gemdertxt.getStyle() + ";-fx-border-color: red;");
+            return;
+        }
+        if (isValidGender&&isvalidAge) {
+
+            AnimalMngDto dto = new AnimalMngDto(animalId, name, age, gender, catogary, encloser, devision);
+            try {
+                boolean add = model.add(dto);
+                if (add) {
+                    new Alert(Alert.AlertType.CONFIRMATION, "Animal Added Successfully").show();
+                    clearform();
+                    tabelload();
+                    nextAnimalId();
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, "Input datas are Wrong").show();
+                e.printStackTrace();
             }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR, "Input datas are Wrong").show();
-            e.printStackTrace();
         }
 
     }
@@ -191,26 +262,34 @@ animaltbl.setItems(AnimalTms);
         agetxt.setText("");
         gemdertxt.setText("");
         catogarytxt.setText("");
-        enclosrtxt.setText("");
-        devisiontxt.setText("");
+        enclosrtxt.getSelectionModel().clearSelection();
+        devisiontxt.getSelectionModel().clearSelection();
     }
 
     @FXML
     void delete(ActionEvent event) {
         AnimalMngDto dto = new AnimalMngDto();
         dto.setAnimalId(idtxt.getText());
-        try {
-            boolean delete = model.delete(dto);
-            if (delete) {
-                new Alert(Alert.AlertType.INFORMATION, "Animal Deleted Successfully").show();
-                tabelload();
-                clearform();
-                nextAnimalId();
-            }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR, "Input datas are Wrong ").show();
-        }
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure?", ButtonType.YES, ButtonType.NO);
+        Optional<ButtonType> optionalButtonType = alert.showAndWait();
 
+        if (optionalButtonType.isPresent() && optionalButtonType.get() == ButtonType.YES) {
+
+            try {
+                boolean delete = model.delete(dto);
+                if (delete) {
+                    new Alert(Alert.AlertType.INFORMATION, "Animal Deleted Successfully").show();
+                    tabelload();
+                    clearform();
+                    nextAnimalId();
+                }
+                else {
+                    new Alert(Alert.AlertType.ERROR, "Animal Not Deleted").show();
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, "Input datas are Wrong ").show();
+            }
+        }
     }
 
     @FXML
@@ -244,8 +323,8 @@ animaltbl.setItems(AnimalTms);
      agetxt.setText(String.valueOf(animalMngTM.getAge()));
      gemdertxt.setText(animalMngTM.getGender());
      catogarytxt.setText(animalMngTM.getCategory());
-     enclosrtxt.setText(animalMngTM.getEnclosureId());
-     devisiontxt.setText(animalMngTM.getDivisionId());
+     enclosrtxt.setValue(animalMngTM.getEnclosureId());
+     devisiontxt.setValue(animalMngTM.getDivisionId());
     }
  }
  public void nextAnimalId() throws SQLException {
