+++ src/main/java/edu/ijjse/dehivalazoomanagemetsystem/controller/EncloserController.java
@@ -5,6 +5,7 @@ import com.jfoenix.controls.JFXTextField;
 import edu.ijjse.dehivalazoomanagemetsystem.dto.EncloserDto;
 import edu.ijjse.dehivalazoomanagemetsystem.dto.tm.EncloserTm;
 import edu.ijjse.dehivalazoomanagemetsystem.model.EncloserModel;
+import edu.ijjse.dehivalazoomanagemetsystem.utill.RegexUtill;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -14,6 +15,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -26,6 +28,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class EncloserController implements Initializable {
@@ -100,20 +103,32 @@ public class EncloserController implements Initializable {
         String location = locationtxt.getText();
         String size = sizetxt.getText();
 
-        EncloserDto encloserDto = new EncloserDto(id,location,size);
-        try {
-            boolean update = model.update(encloserDto);
-            if (update) {
-                new Alert(Alert.AlertType.INFORMATION,"Update Successful").show();
-                loadTbl();
-                clearForm();
+        idtxt.setStyle(idtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        locationtxt.setStyle(locationtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        sizetxt.setStyle(sizetxt.getStyle() + ";-fx-border-color: #7367F0;");
+
+        boolean isValidSize = RegexUtill.IsValidSize(sizetxt.getText());
+        if (!isValidSize) {
+            new Alert(Alert.AlertType.ERROR, "Invalid Size").show();
+            sizetxt.setStyle(sizetxt.getStyle() + ";-fx-border-color: red;");
+        }
+        if (isValidSize) {
+
+
+            EncloserDto encloserDto = new EncloserDto(id, location, size);
+            try {
+                boolean update = model.update(encloserDto);
+                if (update) {
+                    new Alert(Alert.AlertType.INFORMATION, "Update Successful").show();
+                    loadTbl();
+                    clearForm();
+                } else {
+                    new Alert(Alert.AlertType.ERROR, "Update Failed").show();
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
+                e.printStackTrace();
             }
-            else {
-                new Alert(Alert.AlertType.ERROR,"Update Failed").show();
-            }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
-            e.printStackTrace();
         }
 
     }
@@ -124,23 +139,32 @@ public class EncloserController implements Initializable {
         String location = locationtxt.getText();
         String size = sizetxt.getText();
 
-        EncloserDto encloserDto = new EncloserDto(id,location,size);
-        try {
-            boolean add = model.add(encloserDto);
-            if (add) {
-                new Alert(Alert.AlertType.INFORMATION,"Add Successful").show();
-                loadTbl();
-                clearForm();
-            }
-            else {
-                new Alert(Alert.AlertType.ERROR,"Add Failed").show();
-            }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
-            e.printStackTrace();
+        idtxt.setStyle(idtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        locationtxt.setStyle(locationtxt.getStyle() + ";-fx-border-color: #7367F0;");
+        sizetxt.setStyle(sizetxt.getStyle() + ";-fx-border-color: #7367F0;");
+
+        boolean isValidSize = RegexUtill.IsValidSize(sizetxt.getText());
+        if (!isValidSize) {
+            new Alert(Alert.AlertType.ERROR, "Invalid Size").show();
+            sizetxt.setStyle(sizetxt.getStyle() + ";-fx-border-color: red;");
         }
-
-
+         if (isValidSize){
+             EncloserDto encloserDto = new EncloserDto(id,location,size);
+             try {
+                 boolean add = model.add(encloserDto);
+                 if (add) {
+                     new Alert(Alert.AlertType.INFORMATION,"Add Successful").show();
+                     loadTbl();
+                     clearForm();
+                 }
+                 else {
+                     new Alert(Alert.AlertType.ERROR,"Add Failed").show();
+                 }
+             } catch (SQLException e) {
+                 new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
+                 e.printStackTrace();
+             }
+         }
 
     }
     private void clearForm() {
@@ -166,18 +190,23 @@ public class EncloserController implements Initializable {
     void delete(ActionEvent event) {
         EncloserDto encloserDto = new EncloserDto();
         encloserDto.setId(idtxt.getText());
-        try {
-            boolean delete = model.delete(encloserDto);
-            if (delete) {
-                new Alert(Alert.AlertType.INFORMATION,"Delete Successful").show();
-                loadTbl();
-                clearForm();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure?", ButtonType.YES, ButtonType.NO);
+        Optional<ButtonType> optionalButtonType = alert.showAndWait();
+
+        if (optionalButtonType.isPresent() && optionalButtonType.get() == ButtonType.YES) {
+
+            try {
+                boolean delete = model.delete(encloserDto);
+                if (delete) {
+                    new Alert(Alert.AlertType.INFORMATION, "Delete Successful").show();
+                    loadTbl();
+                    clearForm();
+                } else {
+                    new Alert(Alert.AlertType.ERROR, "Delete Failed").show();
+                }
+            } catch (SQLException e) {
+                new Alert(Alert.AlertType.ERROR, e.getMessage()).show();
             }
-            else{
-                new Alert(Alert.AlertType.ERROR,"Delete Failed").show();
-            }
-        } catch (SQLException e) {
-            new Alert(Alert.AlertType.ERROR,e.getMessage()).show();
         }
 
     }
